requirements.txt
asyncpg==0.29.0
fastapi==0.111.0
Jinja2==3.1.4
loguru==0.7.2
passlib==1.7.4
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose==3.3.0
python-multipart==0.0.9
python-ulid==2.7.0
uvicorn==0.30.1
